コンフリクトの解消
