+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'diagnostic_page.dart';
+import 'views/diagnostic_page.dart';
 
 void main() {
   runApp(
@@ -21,13 +21,11 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: Container(
-          child: DiagnosticPage(
-              question: "性格は？",
-              leftText: "活発",
-              rightText: "穏やか",
-              leftOnPressed: (() => {}),
-              rightOnPressed: (() => {}))),
+      home: const DiagnosticPage(
+        question: "性格は？",
+        leftText: "活発",
+        rightText: "穏やか",
+      ),
     );
   }
 }
