+++ lib/ui_components/diagnostic_button.dart
