+++ lib/views/result_page.dart
@@ -1,4 +1,4 @@
-import 'package:biwa_musume_app/providers/diagnosis_notifier.dart';
+import 'package:biwa_musume_app/providers/diagnostic_notifier.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
@@ -9,7 +9,7 @@ class ResultPage extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     return Scaffold(
       body: Center(
-        child: Text('resultIndex: ${ref.read(diagnosisProvider).resultIndex}'),
+        child: Text('resultIndex: ${ref.read(diagnosticProvider).resultIndex}'),
       ),
     );
   }
